added ActorResolver
