+++ src/main/scala/org/sandbox/chat/cluster/ChatClusterActors.scala
@@ -1,22 +1,19 @@
 package org.sandbox.chat.cluster
 
 import scala.annotation.migration
-import scala.concurrent.Await
 import scala.util.Random
-import scala.util.Try
 
 import org.sandbox.chat.cluster.ChatClusterRole.roleToString
 
 import akka.actor.ActorContext
 import akka.actor.ActorRef
-import akka.actor.RootActorPath
 import akka.cluster.Member
 import akka.event.LoggingAdapter
 import akka.util.Timeout
 
 class ChatClusterActors(role: ChatClusterRole,
     implicit val context: ActorContext, implicit val timeout: Timeout,
-    implicit val log: LoggingAdapter)
+    implicit val log: LoggingAdapter) extends ActorResolver
 {
   private var clusterActors: Map[Member,ActorRef] = Map.empty
 
@@ -25,7 +22,7 @@ class ChatClusterActors(role: ChatClusterRole,
 
   def onMemberUp(member: Member) = {
     if (member.hasRole(role)) {
-      val actor = resolveActor(member, role)
+      val actor = resolveActorForMember(member, role)
       actor foreach { a =>
         clusterActors += member -> a
         context watch a
@@ -39,18 +36,4 @@ class ChatClusterActors(role: ChatClusterRole,
 
   def onTerminated(actor: ActorRef) =
     clusterActors = clusterActors.filterNot { case (_, a) => a == actor }
-
-  import context.dispatcher
-
-  private def resolveActor(member: Member, actorName: String): Option[ActorRef] = {
-    val actorPath = RootActorPath(member.address) / "user" / actorName
-    log.info(s"selecting actor $actorPath")
-    val actorSelection = context.actorSelection(actorPath)
-    val actorF = actorSelection.resolveOne
-    actorF onFailure {
-      case e => log.error(s"could not resolve actor $actorPath: ${e.getMessage}")
-    }
-    val actor = Try(Await.result(actorF, timeout.duration))
-    actor.toOption
-  }
 }
